@oscardssmith 2020-12-15
@nanosoldier runbenchmarks(ALL, vs=":master")
Let's find out if I have the perms to do this.