@dkarrasch 2020-12-17
Seems like it didn't work. Let me try:
@nanosoldier `runbenchmarks(ALL, vs=":master")`