@nanosoldier 2020-12-17
[Your benchmark job](https://github.com/JuliaLang/julia/pull/38876#issuecomment-747282085) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/44a6b999a9468b86c3a514ca04e0e3978a440b05/benchmark/by_hash/be17de3_vs_3853060/report.md). cc @christopher-dG